FoodiezBackend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
-e git+https://github.com/kaustuvpokharel/FOODIEZ@8f1eec64241525627910c1ea387c7d9724d67a8c#egg=media_analysis&subdirectory=media_analysis
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
PyJWT==2.9.0
pypdfium2==4.30.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
sympy==1.13.1
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
